# Update Unicode data and optimize CharUnicodeInfo indexes