--- a/src/System.Private.CoreLib/Tools/GenUnicodeProp/Program.cs
+++ b/src/System.Private.CoreLib/Tools/GenUnicodeProp/Program.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.IO;