--- a/src/System.Private.CoreLib/Tools/GenUnicodeProp/Readme.md
+++ b/src/System.Private.CoreLib/Tools/GenUnicodeProp/Readme.md
@@ -1,0 +1,3 @@
+This folder contains the program used to generate the Unicode character categories file CharUnicodeInfoData.cs.  
+The required UnicodeData.txt file can be obtained from https://www.unicode.org/Public/UCD/latest/ucd/UnicodeData.txt  
+To generate the file execute: `dotnet run`