… src/System.Private.CoreLib/shared/System/Globalization/CharUnicodeInfoData.cs