@tarekgh 2018-11-14
> Category data is updated to the latest Unicode 11 data (#20589 somehow missed some changes).

what data did we miss there?