@pentp 2018-11-14
> 
> 
> > Category data is updated to the latest Unicode 11 data (#20589 somehow missed some changes).
> 
> what data did we miss there?

By looking at https://www.unicode.org/Public/UCD/latest/ucd/DerivedAge.txt it looks like the previous data was actually Unicode 10.0