@tarekgh 2018-11-14
Could you please test your changes with the test https://github.com/dotnet/corefx/tree/master/src/System.Globalization/tests from corex master? just to ensure we work fine.