@pentp 2018-11-15
I added big-endian support and tested with CoreFX System.Globalization tests and verified that the test used UnicodeData.11.0.txt as the reference file.